+++ news/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Article
+
+
+admin.site.register(Article)
