+++ news/templates/news_index.html
@@ -8,7 +8,7 @@
     <figure>
         <h1>{{ article.title }}</h1>
         <img src="{{ article.img_source.url }}" height="400" alt="{{ article.title }}">
-        <div>{{ article.get_brief }} <a href="{% url 'pharmacy:news_detail' article.id %}">Continue reading</a></div>
+        <div>{{ article.get_brief }} <a href="{% url 'news:news_detail' article.id %}">Continue reading</a></div>
     </figure>
     {% endfor %}
 </ul>
